requirements.txt
streamlit==1.23.1
pandas==1.4.2
numpy==1.22.4
pillow==9.1.1